+++ app/Http/Controllers/Admin/PelangganController.php
@@ -17,6 +17,33 @@ class PelangganController extends Controller
     {
         $this->middleware('auth:personel');
     }
+
+    public function checkPersonel()
+    {
+        $personel = Auth::guard('personel')->user();
+
+        $jabatan = DB::table('jabatan')
+                        ->join('personel', 'personel.id_jabatan', 'jabatan.id')
+                        ->select(
+                            'jabatan.id', 'jabatan.jabatan', 'personel.id AS personel_id'
+                        )
+                        ->get();
+                        
+        $jb = null;
+        foreach ($jabatan as $j) {
+            if ($j->personel_id == $personel->id) {
+                $jb = $j->jabatan;
+            }
+        }
+
+        if (
+            $jb == 'Chief Executive Officer' ||
+            $jb == 'Chief Operating Officer' ||
+            $jb == 'Chief Technology Officer'
+        ) {
+            return true;
+        } else { return false; }
+    }
     
     /**
      * Display a listing of the resource.
@@ -25,6 +52,10 @@ class PelangganController extends Controller
      */
     public function index()
     {
+        if (!$this->checkPersonel()) {
+            return redirect('/dashboard');
+        }
+        
         $pelanggan = Pelanggan::latest()->get();
         $kota = DB::table('kota')->get();
         $kecamatan = DB::table('kecamatan')->get();
@@ -47,6 +78,10 @@ class PelangganController extends Controller
      */
     public function store(StorePelanggan $request)
     {
+        if (!$this->checkPersonel()) {
+            return redirect('/dashboard');
+        }
+        
         $validated = $request->validated();
         
         Pelanggan::create([
@@ -80,6 +115,10 @@ class PelangganController extends Controller
      */
     public function edit($id)
     {
+        if (!$this->checkPersonel()) {
+            return redirect('/dashboard');
+        }
+        
         $pelanggan = Pelanggan::find($id);
         $kota = DB::table('kota')->get();
         $kecamatan = DB::table('kecamatan')->get();
@@ -103,6 +142,10 @@ class PelangganController extends Controller
      */
     public function update(UpdatePelanggan $request, $id)
     {   
+        if (!$this->checkPersonel()) {
+            return redirect('/dashboard');
+        }
+        
         $validated = $request->validated();
         $pelanggan = Pelanggan::find($id);
         
@@ -140,6 +183,10 @@ class PelangganController extends Controller
      */
     public function destroy($id)
     {
+        if (!$this->checkPersonel()) {
+            return redirect('/dashboard');
+        }
+        
         $pelanggan = Pelanggan::find($id);
         $pelanggan->delete();
